fix: dashboard toolbar layout updates
