+++ public/app/features/dashboard-scene/scene/DashboardLinkRenderer.tsx
@@ -64,7 +64,6 @@ function getStyles(theme: GrafanaTheme2) {
       alignItems: 'center',
       verticalAlign: 'middle',
       marginBottom: theme.spacing(1),
-      marginRight: theme.spacing(1),
     }),
   };
 }
