+++ public/app/features/dashboard-scene/scene/VariableControlsAddButton.tsx
@@ -1,7 +1,9 @@
+import { css } from '@emotion/css';
 import { PointerEventHandler, useCallback } from 'react';
 
+import { GrafanaTheme2 } from '@grafana/data';
 import { Trans } from '@grafana/i18n';
-import { Button } from '@grafana/ui';
+import { Button, useStyles2 } from '@grafana/ui';
 
 import { openAddVariablePane } from '../settings/variables/VariableAddEditableElement';
 import { DashboardInteractions } from '../utils/interactions';
@@ -9,6 +11,7 @@ import { DashboardInteractions } from '../utils/interactions';
 import { DashboardScene } from './DashboardScene';
 
 export function AddVariableButton({ dashboard }: { dashboard: DashboardScene }) {
+  const styles = useStyles2(getStyles);
   const { editview, editPanel, isEditing, viewPanel } = dashboard.useState();
 
   const handlePointerDown: PointerEventHandler = useCallback(
@@ -30,10 +33,22 @@ export function AddVariableButton({ dashboard }: { dashboard: DashboardScene }) 
   }
 
   return (
-    <div className="dashboard-canvas-add-button">
-      <Button icon="plus" variant="primary" fill="text" onPointerDown={handlePointerDown}>
-        <Trans i18nKey="dashboard-scene.variable-controls.add-variable">Add variable</Trans>
-      </Button>
+    <div className={styles.addButton}>
+      <div className="dashboard-canvas-add-button">
+        <Button icon="plus" variant="primary" fill="text" onPointerDown={handlePointerDown}>
+          <Trans i18nKey="dashboard-scene.variable-controls.add-variable">Add variable</Trans>
+        </Button>
+      </div>
     </div>
   );
 }
+
+const getStyles = (theme: GrafanaTheme2) => ({
+  addButton: css({
+    display: 'inline-flex',
+    alignItems: 'center',
+    verticalAlign: 'middle',
+    marginBottom: theme.spacing(1),
+    marginRight: theme.spacing(1),
+  }),
+});
